Moved `InMemoryGenresStore` and `HTTPClientStub` helpers to separate file
